feat: 蓝盾对接权限中心RBAC #7794

+++ src/backend/ci/ext/tencent/process/biz-process-tencent/src/main/kotlin/com/tencent/devops/process/api/service/ServiceTXPipelineResourceImpl.kt
@@ -51,7 +51,7 @@ class ServiceTXPipelineResourceImpl @Autowired constructor(
         return Result(pipelineExportService.exportV2YamlStr(userId, projectId, pipelineId, true))
     }
 
-    override fun getPipelineInfobyId(id: Int): Result<SimplePipeline> {
+    override fun getPipelineInfobyId(id: Long): Result<SimplePipeline> {
         return Result(pipelineListFacadeService.getByAutoIds(listOf(id))[0])
     }
